Addons: Fixing much fo the recent problems with the addons
Some changes on RomHut caused some of this to be updated. Shouldn't
matter to anyone else using manual methods or RomHut, but you will need
to add an <id>0</id> (for example) to every addon in the addons.xml.
This is already done on RomHut. Make sure it's unique per file.

 - Added a handler class for XML parsing of Addons.
 - Changed the Id from with Addons.class to be directly from the XML.
 - Changed the storing of the addons in the Map for cancelling and
updating progress - this fixed udpating the progress and setting the
cancel button to "finished" upon completion
 
 This will have to do for a little bit until I start a rewrite of all
this code.