Major overhaul of the fragments
The lists now open a generic fragment for files. We just pass the details of the file type and the ID of the file to display.

Also added some Vector drawables for post-lollipop devices to use, rather than a lot of images.

Removed bunk SettingsActivity until I'm read for it